feat: notes all-time, quotes widget, nav customization, book ratings
- feat(notes): remove date filter — show all notes always, with search
- feat(dashboard): add QuotesWidget with time-based rotation and prev/next controls
- feat(books): add Quotes panel on every book (save quote + page number, delete)
- feat(books): add 5-star rating — set while finishing or via edit modal
- feat(settings): Navigation section — pick which 4 items fill the bottom nav
- feat(settings): Quote Widget section — set rotation interval (1h / 6h / 12h / 24h)
- feat(store): add navItems + quoteIntervalHours to useAppStore, persisted in localStorage
- fix(types): add rating field to Book type (Row/Insert/Update)
- fix(hooks): update useQuotesQuery to accept optional bookId filter
- fix(hooks): update useQuoteMutations to scope mutations to a bookId
- fix(books): add rating to SUPABASE_BOOK_COLUMNS allowlist
- db: add supabase/migrations/add_rating_to_books.sql — run this in Supabase SQL editor

+++ supabase/migrations/add_rating_to_books.sql
@@ -1,7 +1,7 @@
 -- Migration: add rating column to books table
 -- Run this in your Supabase SQL editor (Dashboard → SQL Editor → New Query)
 
-ALTER TABLE books ADD COLUMN IF NOT EXISTS rating int CHECK (rating >= 1 AND rating <= 5);
+ALTER TABLE books ADD COLUMN IF NOT EXISTS rating smallint CHECK (rating >= 1 AND rating <= 5);
 
 -- Verify
 -- SELECT column_name, data_type FROM information_schema.columns
